src-restructuring
- src folder restructured, decomposed
