+++ deno.json
@@ -9,13 +9,15 @@
     "@deno/dnt": "jsr:@deno/dnt@^0.41.0",
     "@std/assert": "jsr:@std/assert@^0.220.1",
     "@std/expect": "jsr:@std/expect@^1.0.5",
-    "@std/testing": "jsr:@std/testing@^0.220.1"
+    "@std/testing": "jsr:@std/testing@^0.220.1",
+    "#fetch": "./src/fetch/mod.ts",
+    "#response": "./src/response/mod.ts"
   },
   "exports": {
     ".": "./src/mod.ts",
     "./middleware": "./src/middleware/mod.ts",
-    "./response": "./src/response.ts",
-    "./server": "./src/server.ts",
+    "./response": "./src/response/mod.ts",
+    "./server": "./src/server/mod.ts",
     "./url": "./src/url/mod.ts"
   },
   "fmt": {
