+++ src/middleware/log.ts
@@ -1,4 +1,4 @@
-import type { Middleware } from '../types.ts';
+import type { Middleware } from '#fetch';
 
 /** Basic data for log handler. */
 export interface LogData {
